% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{LEIs2df}
\alias{LEIs2df}
\title{Transform given list of legal entities from \code{openLEIs} funtion to a data.frame}
\usage{
LEIs2df(lEntities, wide = T)
}
\arguments{
\item{\code{lEntities}}{list of legal entities as returned by \code{openLEI}.}

\item{\code{wide}}{logical; set \code{TRUE} if you want the data.frame to be reshaped from long to wide.}
}
\value{
data.frame
}
\description{
Transform given list of legal entities from \code{openLEIs} funtion to a data.frame
}
