% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openLEIr.R
\docType{package}
\name{openLEIr-package}
\alias{openLEIr}
\alias{openLEIr-package}
\title{Interface for accessing metadata of financial entities based on their LEI codes from R}
\description{
Interface for accessing metadata of financial entities based on their LEI codes from R
}
